split Client.hs module
